requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
Flask==1.0.3
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.0.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.0
PyJWT==1.7.1
python-dateutil==2.8.0
pyxero==0.9.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4